Add files via upload
This is a repository showing the raw data, the Excel file, a screenshot image and a "ReadMe" file showing an analyzed and visualized data using Excel on "Clean/legal Jobs" in some part of the world over a period of time. It clearly shows the available clean jobs, top five and last five companies most hiring and their location, and other relevant information.